requirements.txt
apparmor==3.0.0
appdirs==1.4.4
application-utility==1.3.2
asn1crypto==1.4.0
btrfsutil==5.9
CacheControl==0.12.6
ceph==1.0.0
ceph-volume==1.0.0
cephfs==2.0.0
cephfs-shell==0.0.1
cffi==1.14.3
chardet==3.0.4
colorama==0.4.4
contextlib2==0.6.0.post1
cryptography==3.2.1
cupshelpers==1.0
distlib==0.3.1
distro==1.5.0
docopt==0.6.2
html5lib==1.1
idna==2.10
keyutils==0.6
LibAppArmor==3.0.0
louis==3.15.0
msgpack==1.0.0
netsnmp-python==1.0a1
npyscreen==4.10.5
ordered-set==4.0.2
packaging==20.4
pacman-mirrors==4.16.4
pep517==0.9.1
Pillow==7.2.0
ply==3.11
progress==1.5
pwquality==1.4.4
pycairo==1.20.0
pycparser==2.20
pycups==2.0.1
pycurl==7.43.0.6
PyGObject==3.38.0
pyOpenSSL==19.1.0
pyparsing==2.4.7
PyQt5==5.15.2
PyQt5-sip==12.8.1
pysmbc==1.0.23
python-libtorrent==1.2.10
PyYAML==5.3.1
rados==2.0.0
rbd==2.0.0
reportlab==3.5.55
requests==2.24.0
resolvelib==0.5.1
retrying==1.3.3
rgw==2.0.0
sip==4.19.24
six==1.15.0
team==1.0
toml==0.10.2
udiskie==2.2.0
urllib3==1.25.10
webencodings==0.5.1
